Add gpioctl command for controlling GPIOs
This works by having a gpioctld service which connects to the Analog Discovery,
and a gpioctl program for instructing it to switch GPIOs on or off.

The GPIO controlling process needs to remain running
since if the handle to the Analog Discovery is released
then it will eventually reset to its default state
which would turn off the power to the Audio Test SoM.

To run gpioctld, for now use:

    systemd-socket-activate --seqpacket --inetd --listen /run/gpioctld.socket gpioctld

systemd services will be forthcoming.