Add debian packaging
This includes dependencies that were omitted from the digilent packages
since it's easier to pull them in via dependencies on this package
than fix the digilent packages.

The recommended way to build this package is by setting up sbuild on debian,
and running the following command from inside the directory.

    sbuild --extra-package ../digilent.adept.runtime_2.16.6-amd64.deb \
           --extra-package ../digilent.waveforms_3.5.4_amd64.deb